Update according to 0.15.4
