+++ kedro-tutorial/kedro_cli.py
@@ -28,14 +28,15 @@
 
 """Command line tools for manipulating a Kedro project.
 Intended to be invoked via `kedro`."""
-
 import os
+import re
 import shutil
 import subprocess
 import sys
 from collections import Counter
 from glob import iglob
 from pathlib import Path
+import webbrowser
 
 import click
 from click import secho, style
@@ -49,16 +50,13 @@ from kedro.cli.utils import (
 )
 from kedro.utils import load_obj
 from kedro.runner import SequentialRunner
-from typing import Iterable, List
+from kedro.context import load_context
+from typing import Iterable, List, Dict
 
 CONTEXT_SETTINGS = dict(help_option_names=["-h", "--help"])
 
 # get our package onto the python path
 PROJ_PATH = Path(__file__).resolve().parent
-sys.path.insert(0, str(PROJ_PATH / "src"))
-os.environ["PYTHONPATH"] = (
-    str(PROJ_PATH / "src") + os.pathsep + os.environ.get("PYTHONPATH", "")
-)
 os.environ["IPYTHONDIR"] = str(PROJ_PATH / ".ipython")
 
 
@@ -76,6 +74,9 @@ TAG_ARG_HELP = """Construct the pipeline using only nodes which have this tag
 attached. Option can be used multiple times, what results in a
 pipeline constructed from nodes having any of those tags."""
 
+PIPELINE_ARG_HELP = """Name of the modular pipeline to run.
+If not set, the project pipeline is run by default."""
+
 ENV_ARG_HELP = """Run the pipeline in a configured environment. If not specified,
 pipeline will run using environment `local`."""
 
@@ -83,11 +84,17 @@ NODE_ARG_HELP = """Run only nodes with specified names."""
 
 FROM_NODES_HELP = """A list of node names which should be used as a starting point."""
 
-TO_NODES_HELP = """A list of node names which should be used as an end point"""
+TO_NODES_HELP = """A list of node names which should be used as an end point."""
+
+FROM_INPUTS_HELP = (
+    """A list of dataset names which should be used as a starting point."""
+)
 
 PARALLEL_ARG_HELP = """Run the pipeline using the `ParallelRunner`.
 If not specified, use the `SequentialRunner`. This flag cannot be used together
 with --runner."""
+
+OPEN_ARG_HELP = """Open the documentation in your default browser after building."""
 
 RUNNER_ARG_HELP = """Specify a runner that you want to run the pipeline with.
 This option cannot be used together with --parallel."""
@@ -98,6 +105,32 @@ including sub-folders."""
 OVERWRITE_HELP = """If Python file already exists for the equivalent notebook,
 overwrite its contents."""
 
+LOAD_VERSION_HELP = """Specify a particular dataset version (timestamp) for loading."""
+
+
+def _split_string(ctx, param, value):
+    return [item for item in value.split(",") if item]
+
+
+def _reformat_load_versions(ctx, param, value) -> Dict[str, str]:
+    """Reformat data structure from tuple to dictionary for `load-version`.
+        E.g ('dataset1:time1', 'dataset2:time2') -> {"dataset1": "time1", "dataset2": "time2"}.
+    """
+    load_version_separator = ":"
+    load_versions_dict = {}
+
+    for load_version in value:
+        load_version_list = load_version.split(load_version_separator)
+        if len(load_version_list) != 2:
+            raise ValueError(
+                "Expected the form of `load_version` to be "
+                "`dataset_name:YYYY-MM-DDThh.mm.ss.sssZ`,"
+                "found {} instead".format(load_version)
+            )
+        load_versions_dict[load_version_list[0]] = load_version_list[1]
+
+    return load_versions_dict
+
 
 @click.group(context_settings=CONTEXT_SETTINGS, name=__file__)
 def cli():
@@ -105,29 +138,44 @@ def cli():
 
 
 @cli.command()
-@click.option("--from-nodes", type=str, default="", help=FROM_NODES_HELP)
-@click.option("--to-nodes", type=str, default="", help=TO_NODES_HELP)
 @click.option(
-    "--node",
-    "-n",
-    "node_names",
-    type=str,
-    default=None,
-    multiple=True,
-    help=NODE_ARG_HELP,
+    "--from-inputs", type=str, default="", help=FROM_INPUTS_HELP, callback=_split_string
 )
+@click.option(
+    "--from-nodes", type=str, default="", help=FROM_NODES_HELP, callback=_split_string
+)
+@click.option(
+    "--to-nodes", type=str, default="", help=TO_NODES_HELP, callback=_split_string
+)
+@click.option("--node", "-n", "node_names", type=str, multiple=True, help=NODE_ARG_HELP)
 @click.option(
     "--runner", "-r", type=str, default=None, multiple=False, help=RUNNER_ARG_HELP
 )
 @click.option("--parallel", "-p", is_flag=True, multiple=False, help=PARALLEL_ARG_HELP)
 @click.option("--env", "-e", type=str, default=None, multiple=False, help=ENV_ARG_HELP)
-@click.option("--tag", "-t", type=str, default=None, multiple=True, help=TAG_ARG_HELP)
-def run(tag, env, parallel, runner, node_names, to_nodes, from_nodes):
+@click.option("--tag", "-t", type=str, multiple=True, help=TAG_ARG_HELP)
+@click.option(
+    "--load-version",
+    "-lv",
+    type=str,
+    multiple=True,
+    help=LOAD_VERSION_HELP,
+    callback=_reformat_load_versions,
+)
+@click.option("--pipeline", type=str, default=None, help=PIPELINE_ARG_HELP)
+def run(
+    tag,
+    env,
+    parallel,
+    runner,
+    node_names,
+    to_nodes,
+    from_nodes,
+    from_inputs,
+    load_version,
+    pipeline,
+):
     """Run the pipeline."""
-    from kedro_tutorial.run import main
-    from_nodes = [n for n in from_nodes.split(",") if n]
-    to_nodes = [n for n in to_nodes.split(",") if n]
-
     if parallel and runner:
         raise KedroCliError(
             "Both --parallel and --runner options cannot be used together. "
@@ -137,13 +185,16 @@ def run(tag, env, parallel, runner, node_names, to_nodes, from_nodes):
         runner = "ParallelRunner"
     runner_class = load_obj(runner, "kedro.runner") if runner else SequentialRunner
 
-    main(
+    context = load_context(Path.cwd(), env=env)
+    context.run(
         tags=tag,
-        env=env,
         runner=runner_class(),
         node_names=node_names,
         from_nodes=from_nodes,
         to_nodes=to_nodes,
+        from_inputs=from_inputs,
+        load_versions=load_version,
+        pipeline_name=pipeline,
     )
 
 
@@ -160,12 +211,19 @@ def test(args):
 
 @cli.command()
 def install():
-    """Install project dependencies from both requirements.txt and environment.yml (optional)."""
+    """Install project dependencies from both requirements.txt
+    and environment.yml (optional)."""
 
     if (Path.cwd() / "src" / "environment.yml").is_file():
         call(["conda", "install", "--file", "src/environment.yml", "--yes"])
 
-    python_call("pip", ["install", "-U", "-r", "src/requirements.txt"])
+    pip_command = ["install", "-U", "-r", "src/requirements.txt"]
+
+    if os.name == "posix":
+        python_call("pip", pip_command)
+    else:
+        command = [sys.executable, "-m", "pip"] + pip_command
+        subprocess.Popen(command, creationflags=subprocess.CREATE_NEW_CONSOLE)
 
 
 @forward_command(cli, forward_help=True)
@@ -184,25 +242,27 @@ def package():
 
 
 @cli.command("build-docs")
-def build_docs():
+@click.option(
+    "--open",
+    "-o",
+    "open_docs",
+    is_flag=True,
+    multiple=False,
+    default=False,
+    help=OPEN_ARG_HELP,
+)
+def build_docs(open_docs):
     """Build the project documentation."""
     python_call("pip", ["install", "src/[docs]"])
     python_call("pip", ["install", "-r", "src/requirements.txt"])
-    python_call(
-        "ipykernel", ["install", "--user", "--name=kedro_tutorial"]
-    )
-    if Path("docs/build").exists():
-        shutil.rmtree("docs/build")
-    call(
-        [
-            "sphinx-apidoc",
-            "--module-first",
-            "-o",
-            "docs/source",
-            "src/kedro_tutorial",
-        ]
-    )
+    python_call("ipykernel", ["install", "--user", "--name=kedro_tutorial"])
+    shutil.rmtree("docs/build", ignore_errors=True)
+    call(["sphinx-apidoc", "--module-first", "-o", "docs/source", "src/kedro_tutorial"])
     call(["sphinx-build", "-M", "html", "docs/source", "docs/build", "-a"])
+    if open_docs:
+        docs_page = (Path.cwd() / "docs" / "build" / "html" / "index.html").as_uri()
+        secho("Opening {}".format(docs_page))
+        webbrowser.open(docs_page)
 
 
 @cli.command("build-reqs")
@@ -259,7 +319,13 @@ def _build_jupyter_command(
     cmd = [base, "--ip=" + ip]
 
     if not all_kernels:
-        cmd.append("--KernelSpecManager.whitelist=['python3']")
+        project_name = "kedro_tutorial"
+        kernel_name = re.sub(r"[^\w]+", "", project_name).strip() or "Kedro"
+
+        cmd += [
+            "--NotebookApp.kernel_spec_manager_class=kedro.cli.jupyter.SingleKernelSpecManager",
+            "--KernelSpecManager.default_kernel_name='{}'".format(kernel_name),
+        ]
 
     return cmd + list(args)
 
@@ -311,25 +377,23 @@ def jupyter_lab(ip, all_kernels, args):
 def convert_notebook(all_flag, overwrite_flag, filepath):
     """Convert selected or all notebooks found in a Kedro project
     to Kedro code, by exporting code from the appropriately-tagged cells:
-
     Cells tagged as `node` will be copied over to a Python file matching
     the name of the notebook, under `src/<package_name>/nodes`.
     *Note*: Make sure your notebooks have unique names!
-
     FILEPATH: Path(s) to exact notebook file(s) to be converted. Both
     relative and absolute paths are accepted.
     Should not be provided if --all flag is already present.
-
     """
-    from kedro_tutorial.run import ProjectContext
+    context = load_context(Path.cwd())
+
     if not filepath and not all_flag:
         secho(
             "Please specify a notebook filepath "
             "or add '--all' to convert all notebooks."
         )
-        return
+        sys.exit(1)
 
-    kedro_project_path = ProjectContext(Path.cwd()).project_path
+    kedro_project_path = context.project_path
     kedro_package_name = "kedro_tutorial"
 
     if all_flag:
