+++ kedro-tutorial/src/kedro_tutorial/pipelines/data_engineering/nodes.py
@@ -25,3 +25,108 @@
 #
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import pandas as pd
+from functools import wraps
+from typing import Callable
+import time
+import logging
+
+
+def log_running_time(func: Callable) -> Callable:
+    """Decorator for logging node execution time.
+
+        Args:
+            func: Function to be executed.
+        Returns:
+            Decorator for logging the running time.
+
+    """
+
+    @wraps(func)
+    def with_time(*args, **kwargs):
+        log = logging.getLogger(__name__)
+        t_start = time.time()
+        result = func(*args, **kwargs)
+        t_end = time.time()
+        elapsed = t_end - t_start
+        log.info("Running %r took %.2f seconds", func.__name__, elapsed)
+        return result
+
+    return with_time
+
+
+def _is_true(x):
+    return x == "t"
+
+
+def _parse_percentage(x):
+    if isinstance(x, str):
+        return float(x.replace("%", "")) / 100
+    return float("NaN")
+
+
+def _parse_money(x):
+    return float(x.replace("$", "").replace(",", ""))
+
+
+@log_running_time
+def preprocess_companies(companies: pd.DataFrame) -> pd.DataFrame:
+    """Preprocess the data for companies.
+
+        Args:
+            companies: Source data.
+        Returns:
+            Preprocessed data.
+
+    """
+
+    companies["iata_approved"] = companies["iata_approved"].apply(_is_true)
+
+    companies["company_rating"] = companies["company_rating"].apply(_parse_percentage)
+
+    return companies
+
+
+@log_running_time
+def preprocess_shuttles(shuttles: pd.DataFrame) -> pd.DataFrame:
+    """Preprocess the data for shuttles.
+
+        Args:
+            shuttles: Source data.
+        Returns:
+            Preprocessed data.
+
+    """
+    shuttles["d_check_complete"] = shuttles["d_check_complete"].apply(_is_true)
+
+    shuttles["moon_clearance_complete"] = shuttles["moon_clearance_complete"].apply(
+        _is_true
+    )
+
+    shuttles["price"] = shuttles["price"].apply(_parse_money)
+
+    return shuttles
+
+
+def create_master_table(
+    shuttles: pd.DataFrame, companies: pd.DataFrame, reviews: pd.DataFrame
+) -> pd.DataFrame:
+    """Combines all data to create a master table.
+
+        Args:
+            shuttles: Preprocessed data for shuttles.
+            companies: Preprocessed data for companies.
+            reviews: Source data for reviews.
+        Returns:
+            Master table.
+
+    """
+    rated_shuttles = shuttles.merge(reviews, left_on="id", right_on="shuttle_id")
+
+    with_companies = rated_shuttles.merge(
+        companies, left_on="company_id", right_on="id"
+    )
+
+    master_table = with_companies.drop(["shuttle_id", "company_id"], axis=1)
+    master_table = master_table.dropna()
+    return master_table
